--- Assets/Scripts/EnemyController.cs
@@ -1,20 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class EnemyController : MonoBehaviour
-{
-    public float speed = 3.5f;
-
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        transform.position += new Vector3(-1, 0, 0) * speed * Time.deltaTime;
-    }
-}
