+++ Assets/UnityChanScripts/EnemyController.cs
@@ -1,0 +1,21 @@
+using UnityEngine;
+
+namespace UnityChanScripts
+{
+    public class EnemyController : MonoBehaviour
+    {
+        public float speed = 3.5f;
+
+        // Start is called before the first frame update
+        void Start()
+        {
+        
+        }
+
+        // Update is called once per frame
+        void Update()
+        {
+            transform.position += new Vector3(-1, 0, 0) * speed * Time.deltaTime;
+        }
+    }
+}
